+++ .storybook/main.js
@@ -17,6 +17,7 @@ module.exports = {
         "../src/widgets/action-bar/*.stories.tsx",
         "../src/widgets/object-view/*.stories.tsx",
         "../src/widgets/field-decoration/*.stories.tsx",
+        "../src/widgets/text-field/*.stories.tsx",
     ],
     addons: ["@storybook/addon-actions", "@storybook/addon-knobs"],
     webpackFinal: async config => {