+++ src/index.ts
@@ -18,7 +18,7 @@ export * from "./widgets/object-view/object-view";
 export * from "./widgets/custom-field/custom-field";
 export * from "./widgets/field-decoration/field-decoration";
 
-export * from "./form/text-field";
+export * from "./widgets/text-field/text-field";
 export * from "./form/password-field";
 export * from "./form/dropdown-field";
 export * from "./form/select-field";